+++ resources/js/components/atoms/WorldMap/constants/MapPoints.ts
