+++ resources/js/components/atoms/WorldMap/types/Styles.ts
