+++ resources/js/components/organisms/home/Feature/FeatureSlide/FeatureSlide.tsx
