+++ resources/js/components/organisms/home/Feature/FeatureSlide/FeatureSlideStyles.tsx
