+++ resources/js/components/organisms/home/Feature/FeatureSlide/SlideButton/index.tsx
@@ -1,0 +1,3 @@
+import SlideButton from "@/components/organisms/home/Feature/FeatureSlide/SlideButton/SlideButton";
+
+export default SlideButton;
