+++ resources/js/components/organisms/home/Feature/FeatureSlide/__tests__/FeatureSlide.test.tsx
