+++ resources/js/components/organisms/home/Feature/FeatureSlide/index.tsx
