+++ resources/js/components/organisms/home/Feature/FeatureSlide/types/Slide.ts
