+++ resources/js/components/organisms/home/Feature/FeatureSlide/types/Style.ts
